Make epmd_propfile a gen_server registering as erl_epmd
Argh! This is so broken...
Some modules call `erl_epmd` directly.
When `epmd_propfile` registers itself as `erl_epmd` it can hijack the
messages sent by the calls to `erl_epmd`.
However, some functions in `erl_epmd` are not implemented as messages
to the server - they're carried out directly in the context of the caller!
E.g. `net_adm:names/0`.
This breaks the whole idea of `-epmd_module` `erl` flag and possibility
of using another EPMD-like system for node discovery without changing
all the bits here and there in the OTP codebase.